Adding files from zlib Madler 1.2.8 for Mac OS X and Win builds
